% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_sunburst_internal.R
\name{as.sunburstDF}
\alias{as.sunburstDF}
\title{Create sunburst DF}
\usage{
as.sunburstDF(DF, value_column = NULL, add_root = TRUE)
}
\arguments{
\item{DF}{dataframe}

\item{value_column}{Quantitative variable}
}
\description{
Create sunburst DF
}
\keyword{internal}
