% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot_internal.R
\name{heatmap_plot_func}
\alias{heatmap_plot_func}
\title{Helper function to get plot}
\usage{
heatmap_plot_func(dat)
}
\arguments{
\item{dat}{dataframe}
}
\description{
Helper function to get plot
}
\keyword{internal}
