% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_struc_internal.R
\name{heatmap_struc_func}
\alias{heatmap_struc_func}
\title{Helper function to get data structured}
\usage{
heatmap_struc_func(dat1)
}
\arguments{
\item{dat1}{dataframe}
}
\description{
Helper function to get data structured
}
\keyword{internal}
