% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_plot_internal.R
\name{plot_sunburst}
\alias{plot_sunburst}
\title{Create the sunburst Plot}
\usage{
plot_sunburst(dat, quant)
}
\arguments{
\item{dat}{dataframe}

\item{quant}{Quantitative variable}
}
\description{
Create the sunburst Plot
}
\keyword{internal}
