% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_plot_internal.R
\name{plot_treemap}
\alias{plot_treemap}
\title{Create the treemap Plot}
\usage{
plot_treemap(dat, quant)
}
\arguments{
\item{dat}{dataframe}

\item{quant}{Quantitative variable}
}
\description{
Create the treemap Plot
}
\keyword{internal}
