% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_plot_internal.R
\name{sankey_plot_func}
\alias{sankey_plot_func}
\title{Create the sankey Plot}
\usage{
sankey_plot_func(dat1, quant)
}
\arguments{
\item{dat1}{dataframe}

\item{quant}{Quantitative variable}
}
\description{
Create the sankey Plot
}
\keyword{internal}
