% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_trelliscope.R
\name{sankey_trelliscope}
\alias{sankey_trelliscope}
\title{Create Sankey Trelliscopes}
\usage{
sankey_trelliscope(
  data,
  facet,
  quant_variable,
  title = "Sankey",
  path = "~sankey"
)
}
\arguments{
\item{data}{dataframe to be turned into sankey trelliscope. Ensure that columns are in the order that the sankey
nodes should be left to right}

\item{facet}{column of variables that you want the trelliscope to group on.}

\item{quant_variable}{column containing the numeric values for the links between nodes.}
}
\value{
Trelliscope Display
}
\description{
Create Sankey Trelliscopes
}
\examples{

sankey_trelliscope(data = dat, facet = "state_company", quant_variable = "amount_award")

}
