% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_trelliscope.R
\name{sunburst_trelliscope}
\alias{sunburst_trelliscope}
\title{Create Sunburst Trelliscopes}
\usage{
sunburst_trelliscope(
  data,
  facet,
  quant_variable,
  title = "Sunburst",
  path = "~Sunburst"
)
}
\arguments{
\item{data}{dataframe to be turned into sankey trelliscope. Ensure that columns are in the order that the sankey
nodes should be left to right}

\item{facet}{column of variables that you want the trelliscope to group on.}

\item{quant_variable}{column containing the numeric values for the links between nodes.}
}
\value{
Trelliscope Display
}
\description{
Create Sunburst Trelliscopes
}
\examples{

sunburst_trelliscope(data = dat, facet = "state_company", quant_variable = "amount_award")

}
